@@ -1,7 +1,8 @@
 package com.predic8.scooter.model;
 
-import javax.persistence.Entity;
-import javax.persistence.Id;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
 
 @Entity
 public class Kunde {
